third_party/WebKit/LayoutTests/media/video-controls-visibility-multimodal-mouse-after-touch.html

<!DOCTYPE html>
<html>
<style>
#no-video-media {
  width: 320px;
  height: 240px;
}
</style>
<script src=video-test.js></script>
<script src=media-file.js></script>
<script src=media-controls.js></script>
<script>
var controls;

function runTest()
{
    video = document.getElementById("no-video-media");

    testExpected("video.paused", true);
    if (!window.testRunner)
        return;

    // Tap (touch input) the play button.
    var coords = mediaControlsButtonCoordinates(video, "play-button");
    eventSender.gestureTapDown(coords[0], coords[1]);
    eventSender.gestureShowPress(coords[0], coords[1]);
    eventSender.gestureTap(coords[0], coords[1]);
    testExpected("video.paused", false);

    // And then hover the control with the mouse.
    eventSender.mouseMoveTo(coords[0], coords[1]);

    // And the controls should remain visible.
    runAfterHideMediaControlsTimerFired(function()
    {
        controls = mediaControlsButton(video, "panel");
        testExpected("getComputedStyle(controls).opacity", 1);
        endTest();
    }, video);
}
</script>
<body>
    <p>Test video element control visibility with multimodal input. The controls
       should remain visible if the last input event was a mouse move over them.</p>
    <p>This test only runs in DRT!</p>

    <video id="no-video-media" controls loop oncanplaythrough="runTest()"></video>
    <script>
        setSrcById("no-video-media", findMediaFile("video", "content/test"));
    </script>
</body>
</html>
